Extract link component to its own file
