+++ packages/react/src/LinkList/index.ts
@@ -1,2 +1,4 @@
-export { LinkList, LinkListLink } from './LinkList'
-export type { LinkListProps, LinkListLinkProps } from './LinkList'
+export { LinkList } from './LinkList'
+export type { LinkListProps } from './LinkList'
+export { LinkListLink } from './LinkListLink'
+export type { LinkListLinkProps } from './LinkListLink'
